+++ CHANGELOG.rst
@@ -2,6 +2,13 @@
 Changelog History
 =================
 
+xskillscore v0.0.12 (2020-01-xx)
+================================
+
+Internal Changes
+----------------
+- ~30-50% speedup for deterministic metrics when ``weights=None``. `Aaron Spring`_
+
 xskillscore v0.0.11 (2020-01-06)
 ================================
 
@@ -16,9 +23,6 @@ Deprecations
 ------------
 - ``mad`` no longer works and is replaced by ``median_absolute_error``. `Riley X. Brady`_
 
-Features
---------
-
 Bug Fixes
 ---------
 - ``skipna`` for ``pearson_r`` and ``spearman_r`` and their p-values now reports accurate results when there are pairwise nans (i.e., nans that occur in different indices in ``a`` and ``b``) `Riley X. Brady`_
@@ -30,3 +34,4 @@ Testing
 - Test that ``skipna=True`` works properly with pairwise nans. `Riley X. Brady`_
 
 .. _`Riley X. Brady`: https://github.com/bradyrx
+.. _`Aaron Spring`: https://github.com/aaronspring